Added RTree.cpp and BTree.cpp
RTree.cpp can currently create an empty R-Tree with a root and specified node capacity. 
BTree.cpp implements a binary search tree (BST) with the ability to insert nodes into the tree and keep the tree balanced.  
I am trying to implement the R-Tree in a similar in a similar way to how I implemented the BST, so that's why I have uploaded BTree.cpp as well. 

RTree.cpp can currently add a root node to the R-tree and add the first element (rectangle) to the  root node.