Placeholder files to show current tests passed
Hi Prof. Chester,

This is just an additional branch that has no relevance to the main branch (since once again, the diff will come out to be the whole file instead of the appropriate changes to files. This branch contains code that is currently passing 25 test cases, all of which work on a tree of height 2. Before the due date tomorrow, I shall try to pass the other test as best as I can and submit the files before due date time.